Fix getUUID return type

+++ medialibrary/src/org/videolan/medialibrary/stubs/StubMedialibrary.java
@@ -302,7 +302,7 @@ public class StubMedialibrary extends Medialibrary {
     }
 
     public Playlist createPlaylist(String name) {
-        Playlist playlist = MLServiceLocator.getAbstractPlaylist(dt.getUUID().longValue(), name, 0);
+        Playlist playlist = MLServiceLocator.getAbstractPlaylist(dt.getUUID(), name, 0);
         dt.mPlaylists.add(playlist);
         onPlaylistsAdded();
         return playlist;
